Move files around and make includes consistent.
